fix height type

+++ src/gameplay/map/renderer/renderers/meshes.rs
@@ -10,7 +10,7 @@ use bevy::{
 pub struct Hexagon3D;
 
 impl Hexagon3D {
-    pub fn create_base(size: f32, starting_angle: f32, height_differences: [f32; 6]) -> Mesh {
+    pub fn create_base(size: f32, starting_angle: f32, height_differences: [u8; 6]) -> Mesh {
         let top_vertices: [[f32; 3]; 6] = [
             get_hex_corner_3d(5, starting_angle, size, &height_differences),
             get_hex_corner_3d(4, starting_angle, size, &height_differences),
